// src/gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  /** A high precision floating point value represented as a string */
  BigFloat: { input: string; output: string };
  /** An arbitrary size integer represented as a string */
  BigInt: { input: string; output: string };
  /** An opaque string using for tracking a position in results during pagination */
  Cursor: { input: any; output: any };
  /** A date wihout time information */
  Date: { input: string; output: string };
  /** A date and time */
  Datetime: { input: string; output: string };
  /** A Javascript Object Notation value serialized as a string */
  JSON: { input: string; output: string };
  /** Any type not handled by the type system */
  Opaque: { input: any; output: any };
  /** A time without date information */
  Time: { input: string; output: string };
  /** A universally unique identifier */
  UUID: { input: string; output: string };
};
// ...
export type Node = {
  /** Retrieves a record by `ID` */
  nodeId: Scalars["ID"]["output"];
};
// ...
export type Journey = Node & {
  __typename?: "journey";
  created_at: Scalars["Datetime"]["output"];
  fare: Scalars["Int"]["output"];
  from_address: Scalars["String"]["output"];
  hubType: Scalars["String"]["output"];
  id: Scalars["UUID"]["output"];
  inbound: Scalars["Boolean"]["output"];
  status: Scalars["String"]["output"];
  /** Globally Unique Record Identifier */
  nodeId: Scalars["ID"]["output"];
  to_address: Scalars["String"]["output"];
  traveller_info: Traveller_Info;
};

export type JourneyConnection = {
  __typename?: "journeyConnection";
  edges: Array<JourneyEdge>;
  pageInfo: PageInfo;
};
// ...
export type Traveller_Info = Node & {
  __typename?: "traveller_info";
  created_at: Scalars["Datetime"]["output"];
  first_name: Scalars["String"]["output"];
  flight_number: Scalars["String"]["output"];
  id: Scalars["UUID"]["output"];
  journeyCollection?: Maybe<JourneyConnection>;
  last_name: Scalars["String"]["output"];
  /** Globally Unique Record Identifier */
  nodeId: Scalars["ID"]["output"];
  passenger_count: Scalars["Int"]["output"];
  phone_number: Scalars["String"]["output"];
};